WIP - Start to include views and serializers related to Asset in v2
