+++ kpi/serializers/v1/ancestor_collections.py
@@ -1,15 +1,10 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 
-from rest_framework import serializers
+from kpi.serializers.v2.ancestor_collections import \
+    AncestorCollectionsSerializer as AncestorCollectionsSerializerV2
 
-from kpi.models import Collection
 
+class AncestorCollectionsSerializer(AncestorCollectionsSerializerV2):
 
-class AncestorCollectionsSerializer(serializers.HyperlinkedModelSerializer):
-    url = serializers.HyperlinkedIdentityField(
-        lookup_field='uid', view_name='collection-detail')
-
-    class Meta:
-        model = Collection
-        fields = ('name', 'uid', 'url')
+    pass
