+++ kpi/serializers/v2/asset.py
@@ -1,0 +1,308 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import
+
+import json
+
+from rest_framework import serializers
+from rest_framework.reverse import reverse
+
+from kpi.fields import RelativePrefixHyperlinkedRelatedField, WritableJSONField, \
+    PaginatedApiField
+from kpi.fields.versioned_hyperlinked_identity import VersionedHyperlinkedIdentityField
+from kpi.models import Asset, AssetVersion, Collection
+from kpi.models.asset import ASSET_TYPES
+from kpi.models.object_permission import get_anonymous_user
+from kpi.utils.url_helper import UrlHelper
+
+from .ancestor_collections import AncestorCollectionsSerializer
+from .asset_version import AssetVersionListSerializer
+from .object_permission import ObjectPermissionNestedSerializer
+
+
+class AssetSerializer(serializers.HyperlinkedModelSerializer):
+    owner = RelativePrefixHyperlinkedRelatedField(
+        view_name='user-detail', lookup_field='username', read_only=True)
+    owner__username = serializers.ReadOnlyField(source='owner.username')
+    url = VersionedHyperlinkedIdentityField(
+        lookup_field='uid', view_name='asset-detail')
+    asset_type = serializers.ChoiceField(choices=ASSET_TYPES)
+    settings = WritableJSONField(required=False, allow_blank=True)
+    content = WritableJSONField(required=False)
+    report_styles = WritableJSONField(required=False)
+    report_custom = WritableJSONField(required=False)
+    map_styles = WritableJSONField(required=False)
+    map_custom = WritableJSONField(required=False)
+    xls_link = serializers.SerializerMethodField()
+    summary = serializers.ReadOnlyField()
+    koboform_link = serializers.SerializerMethodField()
+    xform_link = serializers.SerializerMethodField()
+    version_count = serializers.SerializerMethodField()
+    downloads = serializers.SerializerMethodField()
+    embeds = serializers.SerializerMethodField()
+    parent = RelativePrefixHyperlinkedRelatedField(
+        lookup_field='uid',
+        queryset=Collection.objects.all(),
+        view_name='collection-detail',
+        required=False,
+        allow_null=True
+    )
+    ancestors = AncestorCollectionsSerializer(
+        many=True, read_only=True, source='get_ancestors_or_none')
+    permissions = ObjectPermissionNestedSerializer(many=True, read_only=True)
+    tag_string = serializers.CharField(required=False, allow_blank=True)
+    version_id = serializers.CharField(read_only=True)
+    version__content_hash = serializers.CharField(read_only=True)
+    has_deployment = serializers.ReadOnlyField()
+    deployed_version_id = serializers.SerializerMethodField()
+    deployed_versions = PaginatedApiField(
+        serializer_class=AssetVersionListSerializer,
+        # Higher-than-normal limit since the client doesn't yet know how to
+        # request more than the first page
+        default_limit=100
+    )
+    deployment__identifier = serializers.SerializerMethodField()
+    deployment__active = serializers.SerializerMethodField()
+    deployment__links = serializers.SerializerMethodField()
+    deployment__data_download_links = serializers.SerializerMethodField()
+    deployment__submission_count = serializers.SerializerMethodField()
+
+    # Only add link instead of hooks list to avoid multiple access to DB.
+    hooks_link = serializers.SerializerMethodField()
+
+    class Meta:
+        model = Asset
+        lookup_field = 'uid'
+        fields = ('url',
+                  'owner',
+                  'owner__username',
+                  'parent',
+                  'ancestors',
+                  'settings',
+                  'asset_type',
+                  'date_created',
+                  'summary',
+                  'date_modified',
+                  'version_id',
+                  'version__content_hash',
+                  'version_count',
+                  'has_deployment',
+                  'deployed_version_id',
+                  'deployed_versions',
+                  'deployment__identifier',
+                  'deployment__links',
+                  'deployment__active',
+                  'deployment__data_download_links',
+                  'deployment__submission_count',
+                  'report_styles',
+                  'report_custom',
+                  'map_styles',
+                  'map_custom',
+                  'content',
+                  'downloads',
+                  'embeds',
+                  'koboform_link',
+                  'xform_link',
+                  'hooks_link',
+                  'tag_string',
+                  'uid',
+                  'kind',
+                  'xls_link',
+                  'name',
+                  'permissions',
+                  'settings',)
+        extra_kwargs = {
+            'parent': {
+                'lookup_field': 'uid',
+            },
+            'uid': {
+                'read_only': True,
+            },
+        }
+
+    def update(self, asset, validated_data):
+        asset_content = asset.content
+        _req_data = self.context['request'].data
+        _has_translations = 'translations' in _req_data
+        _has_content = 'content' in _req_data
+        if _has_translations and not _has_content:
+            translations_list = json.loads(_req_data['translations'])
+            try:
+                asset.update_translation_list(translations_list)
+            except ValueError as err:
+                raise serializers.ValidationError(err.message)
+            validated_data['content'] = asset_content
+        return super(AssetSerializer, self).update(asset, validated_data)
+
+    def get_fields(self, *args, **kwargs):
+        fields = super(AssetSerializer, self).get_fields(*args, **kwargs)
+        user = self.context['request'].user
+        # Check if the user is anonymous. The
+        # django.contrib.auth.models.AnonymousUser object doesn't work for
+        # queries.
+        if user.is_anonymous():
+            user = get_anonymous_user()
+        if 'parent' in fields:
+            # TODO: remove this restriction?
+            fields['parent'].queryset = fields['parent'].queryset.filter(
+                owner=user)
+        # Honor requests to exclude fields
+        # TODO: Actually exclude fields from tha database query! DRF grabs
+        # all columns, even ones that are never named in `fields`
+        excludes = self.context['request'].GET.get('exclude', '')
+        for exclude in excludes.split(','):
+            exclude = exclude.strip()
+            if exclude in fields:
+                fields.pop(exclude)
+        return fields
+
+    def get_version_count(self, obj):
+        return obj.asset_versions.count()
+
+    def get_xls_link(self, obj):
+        return UrlHelper.reverse('asset-xls',
+                                 args=(obj.uid,),
+                                 request=self.context.get('request', None),
+                                 context=self.context)
+
+    def get_xform_link(self, obj):
+        return UrlHelper.reverse('asset-xform',
+                                 args=(obj.uid,),
+                                 request=self.context.get('request', None),
+                                 context=self.context)
+
+    def get_hooks_link(self, obj):
+        return UrlHelper.reverse('hook-list',
+                                 args=(obj.uid,),
+                                 request=self.context.get('request', None),
+                                 context=self.context)
+
+    def get_embeds(self, obj):
+        request = self.context.get('request', None)
+
+        def _reverse_lookup_format(fmt):
+            url = UrlHelper.reverse('asset-%s' % fmt,
+                                    args=(obj.uid,),
+                                    request=request)
+            return {'format': fmt,
+                    'url': url, }
+
+        return [
+            _reverse_lookup_format('xls'),
+            _reverse_lookup_format('xform'),
+        ]
+
+    def get_downloads(self, obj):
+        def _reverse_lookup_format(fmt):
+            request = self.context.get('request', None)
+            obj_url = UrlHelper.reverse('asset-detail',
+                                        args=(obj.uid,),
+                                        request=request,
+                                        context=self.context)
+            # The trailing slash must be removed prior to appending the format
+            # extension
+            url = '%s.%s' % (obj_url.rstrip('/'), fmt)
+
+            return {'format': fmt,
+                    'url': url, }
+        return [
+            _reverse_lookup_format('xls'),
+            _reverse_lookup_format('xml'),
+        ]
+
+    def get_koboform_link(self, obj):
+        return UrlHelper.reverse('asset-koboform',
+                                 args=(obj.uid,),
+                                 request=self.context.get('request', None),
+                                 context=self.context)
+
+    def get_deployed_version_id(self, obj):
+        if not obj.has_deployment:
+            return
+        if isinstance(obj.deployment.version_id, int):
+            asset_versions_uids_only = obj.asset_versions.only('uid')
+            # this can be removed once the 'replace_deployment_ids'
+            # migration has been run
+            v_id = obj.deployment.version_id
+            try:
+                return asset_versions_uids_only.get(
+                    _reversion_version_id=v_id
+                ).uid
+            except AssetVersion.DoesNotExist:
+                deployed_version = asset_versions_uids_only.filter(
+                    deployed=True
+                ).first()
+                if deployed_version:
+                    return deployed_version.uid
+                else:
+                    return None
+        else:
+            return obj.deployment.version_id
+
+    def get_deployment__identifier(self, obj):
+        if obj.has_deployment:
+            return obj.deployment.identifier
+
+    def get_deployment__active(self, obj):
+        return obj.has_deployment and obj.deployment.active
+
+    def get_deployment__links(self, obj):
+        if obj.has_deployment and obj.deployment.active:
+            return obj.deployment.get_enketo_survey_links()
+        else:
+            return {}
+
+    def get_deployment__data_download_links(self, obj):
+        if obj.has_deployment:
+            return obj.deployment.get_data_download_links()
+        else:
+            return {}
+
+    def get_deployment__submission_count(self, obj):
+        if not obj.has_deployment:
+            return 0
+        return obj.deployment.submission_count
+
+    def _content(self, obj):
+        return json.dumps(obj.content)
+
+    def _table_url(self, obj):
+        request = self.context.get('request', None)
+        return UrlHelper.reverse('asset-table-view',
+                                 args=(obj.uid,),
+                                 request=request,
+                                 context=self.context)
+
+
+class AssetListSerializer(AssetSerializer):
+    class Meta(AssetSerializer.Meta):
+        # WARNING! If you're changing something here, please update
+        # `Asset.optimize_queryset_for_list()`; otherwise, you'll cause an
+        # additional database query for each asset in the list.
+        fields = ('url',
+                  'date_modified',
+                  'date_created',
+                  'owner',
+                  'summary',
+                  'owner__username',
+                  'parent',
+                  'uid',
+                  'tag_string',
+                  'settings',
+                  'kind',
+                  'name',
+                  'asset_type',
+                  'version_id',
+                  'has_deployment',
+                  'deployed_version_id',
+                  'deployment__identifier',
+                  'deployment__active',
+                  'deployment__submission_count',
+                  'permissions',
+                  'downloads',
+                  )
+
+
+class AssetUrlListSerializer(AssetSerializer):
+
+    class Meta(AssetSerializer.Meta):
+        fields = ('url',)
