+++ kpi/urls/router_api_v2.py
@@ -5,13 +5,16 @@ from rest_framework_extensions.routers import ExtendedDefaultRouter
 
 from kobo.apps.hook.views import HookViewSet, HookLogViewSet
 from kpi.views.v1 import (
-    AssetViewSet,
     AssetVersionViewSet,
-    AssetSnapshotViewSet,
     AssetFileViewSet,
     HookSignalViewSet,
     SubmissionViewSet,
 )
+
+from kpi.views.v2.asset import AssetViewSet
+from kpi.views.v2.collection import CollectionViewSet
+from kpi.views.v2.asset_snapshot import AssetSnapshotViewSet
+from kpi.views.v2.user import UserViewSet
 
 
 router_api_v2 = ExtendedDefaultRouter()
@@ -26,7 +29,7 @@ asset_routes.register(r'hook-signal',
                       base_name='hook-signal',
                       parents_query_lookups=['asset'],
                       )
-asset_routes.register(r'submissions',
+asset_routes.register(r'data',
                       SubmissionViewSet,
                       base_name='submission',
                       parents_query_lookups=['asset'],
@@ -50,3 +53,5 @@ hook_routes.register(r'logs',
                      )
 
 router_api_v2.register(r'asset_snapshots', AssetSnapshotViewSet)
+router_api_v2.register(r'collections', CollectionViewSet)
+router_api_v2.register(r'users', UserViewSet)
